gen.py: provide richer contextual information for annotating the text
When supplying context for output,
include the start position and length of the contextual output.

This information can be translated to JSON and can be parsed into a
program that can render the text by annotating the contextual text
rather than simply adding cues such as [0], [1], etc.

The first and only example as of this moment will inject the
contextual cues as [0], [1], ... just as before, but now a little
bit tricker.

A potential use is to re-render HTML links from the position and the
supplied href link.